requirements.txt
py-jwt
requests
fastapi
python-gnupg
sqlalchemy
uvicorn
jinja
asyncio